requirements.txt
# Зависимости проекта: Улучшение базовой модели недвижимости
# Автор: Щербаков Дмитрий
# Объединено для всех этапов проекта

# ===== ОСНОВНЫЕ ML БИБЛИОТЕКИ =====
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.4.1.post1
matplotlib==3.8.3
seaborn>=0.12.0
scipy==1.12.0

# ===== МОДЕЛИ МАШИННОГО ОБУЧЕНИЯ =====
catboost==1.2.3
xgboost==2.0.0
lightgbm>=4.0.0

# ===== MLFLOW И ОТСЛЕЖИВАНИЕ ЭКСПЕРИМЕНТОВ =====
mlflow==2.7.1
sqlalchemy==1.4.50
alembic==1.13.0
psycopg2-binary==2.9.9

# ===== ОПТИМИЗАЦИЯ ГИПЕРПАРАМЕТРОВ =====
optuna>=3.3.0
hyperopt>=0.2.7

# ===== ПРОДВИНУТЫЙ ОТБОР ПРИЗНАКОВ =====
mlxtend>=0.22.0
feature-engine>=1.6.0

# ===== АВТОМАТИЧЕСКАЯ ГЕНЕРАЦИЯ ПРИЗНАКОВ =====
autofeat>=2.0.0

# ===== РАБОТА С ОКРУЖЕНИЕМ И КОНФИГУРАЦИЕЙ =====
python-dotenv==1.0.1
pyyaml==6.0.1
configparser>=5.3.0

# ===== ВИЗУАЛИЗАЦИЯ И ОТЧЕТЫ =====
plotly==5.19.0
kaleido>=0.2.1
wordcloud>=1.9.0

# ===== JUPYTER И ИНТЕРАКТИВНАЯ РАЗРАБОТКА =====
jupyter>=1.0.0
ipykernel==6.29.3
ipywidgets>=8.1.0
notebook>=7.0.0

# ===== ВЕБ ФРЕЙМВОРКИ (ДЛЯ РАЗВЕРТЫВАНИЯ) =====
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic==2.5.2

# ===== РАБОТА С ФАЙЛАМИ И ДАННЫМИ =====
openpyxl>=3.1.0
xlrd>=2.0.0
requests==2.31.0

# ===== УТИЛИТЫ И ВСПОМОГАТЕЛЬНЫЕ БИБЛИОТЕКИ =====
tqdm==4.66.2
joblib==1.3.2
pickle5>=0.0.12
cloudpickle>=2.2.0

# ===== ТЕСТИРОВАНИЕ И КАЧЕСТВО КОДА =====
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# ===== РАБОТА С AWS/S3 =====
boto3==1.34.51
botocore==1.34.51

# ===== СТАТИСТИКА И АНАЛИЗ =====
statsmodels==0.14.1
pingouin>=0.5.3

# ===== БЕЗОПАСНОСТЬ И ВАЛИДАЦИЯ =====
validators>=0.21.0
cryptography==41.0.7

# ===== ПРОФИЛИРОВАНИЕ И МОНИТОРИНГ =====
memory-profiler>=0.61.0
psutil==5.9.6

# ===== СОВМЕСТИМОСТЬ И СИСТЕМА =====
packaging==23.2
setuptools>=68.0.0
wheel>=0.41.0

# ===== ДОПОЛНИТЕЛЬНЫЕ УТИЛИТЫ =====
rich==13.7.0          # Красивый вывод в терминале
click==8.1.7           # CLI интерфейсы
pathlib2>=2.3.7       # Работа с путями (для совместимости)
typing_extensions==4.8.0

# ===== AIRFLOW И DVC (ИНФРАСТРУКТУРА) =====
apache-airflow==2.7.3
dvc==3.48.1
hydra-core==1.3.2


# ===== УСТАНОВКА =====
# pip install -r requirements.txt